requirements.txt
requests
djangorestframework
ipython_genutils
Django==2.1
pytz==2019.2
